Dynamically load list of allowed cities
Although the modules for all cities were already being loaded
dynamically before, the list of supported cities wasn't. Now it is. Not
a big fan of the global variable, but it's indexing a list of files when
the server starts and then storing that list of strings in RAM, so it's
probably better this way.

See #18