Scrape data to DB and serve it from there
Scraper now talks to a Postgres DB where it saves the data.
The server does the same and just takes the last set of data
it finds to return that to the user. No more live scraping at
all. Nice.

This of course also means that the scraper will have to be run
periodically (cron, systemdtimer, whatever) for new data to be
accessible. But nothing is lost anymore or saved as a simple
file to disk.

Also changed some stuff in the config file and how the dev
environment variable is handled. If it's set everything is
currently done in the old way. If not, the DB is used instead.

See #54, #10